+++ core/src/lib.rs
@@ -50,6 +50,7 @@ pub mod rpc_subscriptions;
 pub mod sendmmsg;
 pub mod service;
 pub mod sigverify;
+pub mod sigverify_shreds;
 pub mod sigverify_stage;
 pub mod snapshot_packager_service;
 pub mod storage_stage;
